Reuse original env in preview request
Instead of using the original request’s updated env after the request completes, we save a copy of it as it was when we first received it. This, combined with moving the rehearsal middleware up the stack as far as we can (right after Rack::Lock) gives us a much more pristine env hash with fewer keys to populate ourselves.