Don’t inject middleware at a specific position
It turns out `Rack::Lock` isn’t always present. Since we’re now saving a copy of the original request env before it hits Rails, we should get a fairly clean env hash for the second request. If necessary the host app can move the middleware to the desired location, but by default it will be appended.