Prevent form params from being passed on to preview
Rack includes a separate header for form params that we were not removing. If not removed, these params will be passed on to the preview request where they will be included in the params hash and could affect behaviour.

This was observed in a request that updated a page, and then showed the index page. The update stored page attributes in a ‘page’ param, and the preview request looked in the `page` param for pagination information.